commit_analyser/requirements.txt
GitPython==3.1.24
pyyaml==6.0
